# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Wisetsuk Nakorn Condo Booking API
  version: 1.0.0
  description: |
    REST API ระบบจองห้องพักวิเศษสุขนครคอนโด
    แหล่งความจริงเดียวของ contract — ทุก endpoint ตรงกับหัวข้อ API Contract ใน docs/SPEC.md
servers:
  - url: http://localhost:8080
    description: development
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/bookings/{bookingID}/payment:
    post:
      tags: [bookings]
      summary: แจ้งชำระเงินพร้อมแนบสลิป
      operationId: submitPayment
      parameters:
        - $ref: '#/components/parameters/BookingIDPath'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [slip, amount, transferred_at]
              properties:
                slip:
                  type: string
                  format: binary
                amount:
                  type: number
                transferred_at:
                  type: string
                  format: date-time
                note:
                  type: string
      responses:
        '200':
          description: แจ้งชำระเงินสำเร็จ สถานะเป็น awaiting_review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400':
          description: ไฟล์ไม่ถูกต้องหรือใหญ่เกินกำหนด
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: bad_request
                  message: ไฟล์สลิปไม่ถูกต้อง
        '409':
          $ref: '#/components/responses/InvalidState'
        '422':
          $ref: '#/components/responses/ValidationFailed'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    BookingIDPath:
      name: bookingID
      in: path
      required: true
      description: id ของการจอง (bkg-001) ไม่ใช่ booking_code
      schema:
        type: string
        pattern: "^bkg-[0-9]+$"
...
  responses:
...
    InvalidState:
      description: สถานะปัจจุบันทำรายการนี้ไม่ได้
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: invalid_state
              message: สถานะปัจจุบันทำรายการนี้ไม่ได้
    ValidationFailed:
      description: ข้อมูลไม่ผ่านการตรวจสอบ
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: validation_failed
              message: ข้อมูลไม่ถูกต้อง
              fields:
                email: อีเมลนี้ถูกใช้งานแล้ว
  schemas:
    ErrorCode:
      type: string
      enum:
        - unauthorized
        - forbidden
        - not_found
        - conflict
        - bad_request
        - validation_failed
        - method_not_allowed
        - internal_error
        - invalid_credentials
        - account_disabled
        - room_unavailable
        - invalid_state
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              $ref: '#/components/schemas/ErrorCode'
            message:
              type: string
              description: ข้อความภาษาไทยสำหรับผู้ใช้
            fields:
              type: object
              nullable: true
              description: มีเฉพาะตอน code = validation_failed
              additionalProperties:
                type: string
    StayType:
      type: string
      enum: [daily, monthly]
    BookingStatus:
      type: string
      enum: [pending_payment, awaiting_review, approved, cancelled]
...
    PaymentStatus:
      type: string
      enum: [pending, approved, rejected]
...
    Booking:
      type: object
      required: [id, booking_code, room_id, branch_id, user_id, stay_type, status, total_amount, created_at]
      properties:
        id:
          type: string
          pattern: "^bkg-[0-9]+$"
        booking_code:
          type: string
          pattern: '^PT-[0-9]{6,}$'
          description: ใช้แสดงผลและค้นหาเท่านั้น ห้ามใช้อ้างอิงใน URL
        room_id:
          type: string
          pattern: "^rm-[0-9]+$"
        branch_id:
          type: string
          pattern: "^brn-[0-9]+$"
        user_id:
          type: string
          pattern: "^usr-[0-9]+$"
        stay_type:
          $ref: '#/components/schemas/StayType'
        status:
          $ref: '#/components/schemas/BookingStatus'
        guest_first_name:
          type: string
        guest_last_name:
          type: string
        guest_phone:
          type: string
        check_in_date:
          type: string
          format: date
          nullable: true
        check_out_date:
          type: string
          format: date
          nullable: true
        move_in_date:
          type: string
          format: date
          nullable: true
        nights:
          type: integer
          nullable: true
          description: check_out_date - check_in_date เมื่อ stay_type = daily
        price_per_night:
          type: number
          nullable: true
          description: ราคาที่คัดลอกไว้ตอนจอง ไม่เปลี่ยนตามราคาห้องภายหลัง
        contract_fee:
          type: number
          nullable: true
          description: ค่าทำสัญญาที่คัดลอกไว้ตอนจอง
        total_amount:
          type: number
        expires_at:
          type: string
          format: date-time
          nullable: true
          description: มีค่าเมื่อสถานะเป็น pending_payment
        appointment_at:
          type: string
          format: date-time
          nullable: true
        appointment_note:
          type: string
          nullable: true
        payments:
          type: array
          description: ประวัติการแจ้งชำระเงินทุกครั้ง เรียงจากเก่าไปใหม่
          items:
            $ref: '#/components/schemas/PaymentAttempt'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PaymentAttempt:
      type: object
      required: [id, amount, transferred_at, status, created_at]
      properties:
        id:
          type: string
          pattern: "^pay-[0-9]+$"
        amount:
          type: number
          description: ยอดที่ผู้ใช้แจ้ง อาจไม่ตรงกับ total_amount ให้ผู้ดูแลเป็นผู้ตัดสิน
        transferred_at:
          type: string
          format: date-time
        note:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/PaymentStatus'
        reason:
          type: string
          nullable: true
          description: เหตุผลที่ปฏิเสธ มีค่าเมื่อ status = rejected
        created_at:
          type: string
          format: date-time
...